# src/main/java/com/vehicle/registration/dto/PersonRequest.java
package com.vehicle.registration.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class PersonRequest {
    private String firstName;
    private String lastName;
}