# src/main/java/com/vehicle/registration/dto/PersonVehicleResponse.java
package com.vehicle.registration.dto;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.Data;
import lombok.experimental.SuperBuilder;

@SuperBuilder
@Data
@JsonIgnoreProperties(ignoreUnknown = true)
@JsonInclude(JsonInclude.Include.NON_NULL)
public class PersonVehicleResponse extends AbstractResponse{
    private PersonVehicles personVehicles;
}